feat: Add Clustering module with KMeansDetector class
This commit adds a new Clustering module to the Detection directory. This module contains a new breakpoint detection class called KMeansDetector, also known as CamisDetector, which uses a K-means clustering-based approach for breakpoint detection.

In addition, we added unit tests for the new class to ensure its methods are working as expected. These tests are located in the test_detection.py file in the tests directory.

Lastly, we created a Jupyter notebook called Comparison.ipynb that contains code for comparing the performance of CamisDetector with other breakpoint detection methods in the Ruptures library. This notebook is located in the tests directory.

Detailed comments and DocString documentation were added to explain the functioning of each method in the KMeansDetector class.